requirements.txt
Django==2.2.1
djangorestframework==3.9.3
mysqlclient
bcrypt==3.1.6

